requirements.txt
numpy
soundfile
librosa
einops
torch
pytorch_lightning
hydra-core
torchaudio
